분리된 RewardColor 기능 구현 및 Reward.cs 수정
