+++ Assets/Scripts/RewardColor.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: f219719674fe6a24bad8c612c84e3f99
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
